volplugin: Move to a standalone binary
