--- volplugin/cli.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/codegangsta/cli"
-)
-
-func main() {
-	app := cli.NewApp()
-	app.Name = "volplugin"
-	app.Version = "0.0.1"
-	app.Flags = []cli.Flag{
-		cli.BoolFlag{
-			Name:   "debug",
-			Usage:  "Turn on debug logging",
-			EnvVar: "DEBUG",
-		},
-	}
-	app.Action = daemon
-
-	app.Run(os.Args)
-}
