+++ volplugin/handlers.go
@@ -1,4 +1,4 @@
-package main
+package volplugin
 
 import (
 	"encoding/json"
