+++ volplugin/utils.go
@@ -1,4 +1,4 @@
-package main
+package volplugin
 
 import (
 	"bytes"
